under_edition/requirements_outdated.txt
scipy==1.9.1
pandas==1.4.4
pyarrow==8.0.0
seaborn==0.11.2
scikit-learn==1.3.2
python-dateutil==2.8.2

hyperopt==0.2.7
isodate==0.6.1
ipykernel==6.29.4

lightgbm==3.3.4
numpy==1.22
openpyxl==3.1.2
optuna==3.5.0
pipdeptree==2.13.2  
plotly ==5.20.0
kaleido==0.2.1

researchpy==0.3.5  

py4j==0.10.9.5
pyYAML==6.0.1
#requests-oauthlib==1.3.1

shap==0.44.1
xgboost==2.0.3
XlsxWriter==3.2.0

termcolor==2.4.0

#jupyterlab>=3 
#ipywidgets>=7.6
#jupyter-dash==0.4.2

#ptitprince==0.2.7
termcolor==2.4.0
nbformat==5.10.3
yellowbrick==1.5

psycopg2==2.9.9
pyodbc==5.1.0
SQLAlchemy==2.0.30
jupyter==1.1.1

spacy==3.7.5
requests==2.27.1

#echo yes Y | conda create --name ml_env python=3.8
#conda activate ml_env
#pip install -r  requirments.txt --user
#python -m ipykernel install --user --name ml_env --display-name "ml_env"

##python -m spacy download en_core_web_sm doesnot work see:
##https://stackoverflow.com/questions/55742788/ssl-certificate-verify-failed-error-while-downloading-python-m-spacy-download
##https://spacy.io/models/en#en_core_web_sm
###pip install --target C:/users/rnourzadeh/appdata/local/miniconda3/envs/ml_env/lib/site-packages/spacy/ ./en_core_web_sm-3.7.1-py3-none-any.whl

